feat(security): implement CSP nonce handling and update views for compliance #167
